Spacebux data persistence and updating number in UI, temp. disabled population/power, network and resource bar prefabs
You NEED to start from Login scene if you want to test persistence
because that's where the load code is.
You also have to click on QUIT when exiting because that is where the
save code is.

+++ Assets/Scripts/Spaceobjects/Resource.cs
@@ -8,7 +8,7 @@ public class Resource : MonoBehaviour {
     public float capacity;
 
     //Variables accessible by subclass
-    protected float _amountIncrease;
+    protected uint _amountIncrease;
     protected Utility.ResourceType _resourceType;
     protected Planet _planet;
 
@@ -18,7 +18,7 @@ public class Resource : MonoBehaviour {
     //These methods are virtual and thus can be overriden in child classes
     protected virtual void Start() {
         capacity = 0f;
-        _amountIncrease = 0f;
+        _amountIncrease = 0;
         _resourceType = Utility.ResourceType.Unknown;
     }
